изменен конфиг webpack
