+++ webpack.config.js
@@ -2,6 +2,7 @@ const isProduction = process.env.NODE_ENV === 'production';
 
 module.exports = {
     mode: process.env.NODE_ENV,
+    devtool: isProduction ? false : 'inline-source-map',
     module: {
         rules: [
             {
